requirements.txt
boto3
py-zabbix
